--- project/SantanderSample/SantanderSample/Detail/DetailViewController.swift
@@ -1,150 +1,0 @@
-//
-//  DetailViewController.swift
-//  SantanderSample
-//
-//  Created by Virgilius Santos on 26/10/18.
-//  Copyright (c) 2018 Virgilius Santos. All rights reserved.
-//
-//  This file was generated by the Clean Swift Xcode Templates so
-//  you can apply clean architecture to your iOS and Mac projects,
-//  see http://clean-swift.com
-//
-
-import UIKit
-
-protocol DetailDisplayLogic: class
-{
-  func displaySomething(viewModel: Detail.Something.ViewModel)
-}
-
-class DetailViewController: UIViewController, DetailDisplayLogic
-{
-    
-    @IBOutlet weak var exitView: ExitButtonView! {
-        didSet {
-            let button = exitView.exitButton
-            button?.addTarget(self,
-                              action: #selector(exitAction),
-                              for: .touchUpInside)
-        }
-    }
-    
-    @IBOutlet weak var nameView: InfoDetailView!
-    
-    @IBOutlet weak var accountTitleView: TitleDetailView! {
-        didSet {
-            accountTitleView.titleLabel.text = "Conta"
-        }
-    }
-    @IBOutlet weak var accountInfoView: InfoDetailView!
-    
-    @IBOutlet weak var balanceTitleView: TitleDetailView! {
-        didSet {
-            balanceTitleView.titleLabel.text = "Saldo"
-        }
-    }
-    @IBOutlet weak var balnceInfoView: InfoDetailView!
-    
-    @IBOutlet weak var entriesCollectionView: UICollectionView! {
-        didSet {
-            entriesCollectionView.delegate = self
-            entriesCollectionView.dataSource = self
-            let nib = UINib(nibName: cellIdentifier, bundle: nil)
-            entriesCollectionView
-                .register(nib, forCellWithReuseIdentifier: cellIdentifier)
-        }
-    }
-    
-    
-    let cellIdentifier = String(describing: DetailCell.self)
-    
-    var interactor: DetailBusinessLogic?
-    var router: (NSObjectProtocol & DetailRoutingLogic & DetailDataPassing)?
-
-    // MARK: Object lifecycle
-
-    override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: Bundle?)
-    {
-        super.init(nibName: nibNameOrNil, bundle: nibBundleOrNil)
-        setup()
-    }
-  
-    required init?(coder aDecoder: NSCoder)
-    {
-        super.init(coder: aDecoder)
-        setup()
-    }
-  
-  // MARK: Setup
-  
-  private func setup()
-  {
-    let viewController = self
-    let interactor = DetailInteractor()
-    let presenter = DetailPresenter()
-    let router = DetailRouter()
-    viewController.interactor = interactor
-    viewController.router = router
-    interactor.presenter = presenter
-    presenter.viewController = viewController
-    router.viewController = viewController
-    router.dataStore = interactor
-  }
-  
-  // MARK: Routing
-  
-  override func prepare(for segue: UIStoryboardSegue, sender: Any?)
-  {
-    if let scene = segue.identifier {
-      let selector = NSSelectorFromString("routeTo\(scene)WithSegue:")
-      if let router = router, router.responds(to: selector) {
-        router.perform(selector, with: segue)
-      }
-    }
-  }
-  
-  // MARK: View lifecycle
-  
-  override func viewDidLoad()
-  {
-    super.viewDidLoad()
-    doSomething()
-  }
-  
-  // MARK: Do something
-  
-  //@IBOutlet weak var nameTextField: UITextField!
-  
-  func doSomething()
-  {
-    let request = Detail.Something.Request()
-    interactor?.doSomething(request: request)
-  }
-  
-  func displaySomething(viewModel: Detail.Something.ViewModel)
-  {
-    //nameTextField.text = viewModel.name
-  }
-    
-    @objc func exitAction() {
-        
-    }
-}
-
-extension DetailViewController: UICollectionViewDelegate {
-    
-}
-
-extension DetailViewController: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return 0
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: cellIdentifier, for: indexPath) as! DetailCell
-        
-        return cell
-    }
-    
-    
-}
